requirements.txt
celery==4.3.0
Django==2.2.4
django-import-export==1.2.0
django-restframework-apiview==1.3.15
djangorestframework==3.10.2
djangorestframework-csv==2.1.0
httpie==1.0.3
